requirements.txt
aws-cdk-lib==2.181.1
constructs>=10.0.0,<11.0.0
boto3>=1.40
cdklabs.generative-ai-cdk-constructs
PyYAML>=6.0